refactor(Menu): move production code for menu
