+++ src/main/java/kitchenpos/menu/application/MenuGroupService.java
@@ -1,7 +1,7 @@
-package kitchenpos.asis.application;
+package kitchenpos.menu.application;
 
-import kitchenpos.asis.domain.MenuGroup;
-import kitchenpos.asis.domain.MenuGroupRepository;
+import kitchenpos.menu.domain.MenuGroup;
+import kitchenpos.menu.domain.MenuGroupRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
