+++ src/main/java/kitchenpos/menu/domain/Menu.java
@@ -1,4 +1,4 @@
-package kitchenpos.asis.domain;
+package kitchenpos.menu.domain;
 
 import jakarta.persistence.CascadeType;
 import jakarta.persistence.Column;
