+++ src/main/java/kitchenpos/menu/infra/JpaMenuRepository.java
@@ -1,5 +1,7 @@
-package kitchenpos.asis.domain;
+package kitchenpos.menu.infra;
 
+import kitchenpos.menu.domain.Menu;
+import kitchenpos.menu.domain.MenuRepository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
