+++ src/main/java/kitchenpos/product/application/ProductService.java
@@ -1,8 +1,8 @@
 package kitchenpos.product.application;
 
-import kitchenpos.asis.domain.Menu;
-import kitchenpos.asis.domain.MenuProduct;
-import kitchenpos.asis.domain.MenuRepository;
+import kitchenpos.menu.domain.Menu;
+import kitchenpos.menu.domain.MenuProduct;
+import kitchenpos.menu.domain.MenuRepository;
 import kitchenpos.product.domain.Product;
 import kitchenpos.product.domain.ProductRepository;
 import kitchenpos.asis.infra.PurgomalumClient;
