+++ src/test/java/kitchenpos/Fixtures.java
@@ -1,71 +1,20 @@
 package kitchenpos;
 
-import kitchenpos.asis.domain.Menu;
-import kitchenpos.asis.domain.MenuGroup;
-import kitchenpos.asis.domain.MenuProduct;
 import kitchenpos.asis.domain.Order;
 import kitchenpos.asis.domain.OrderLineItem;
 import kitchenpos.asis.domain.OrderStatus;
 import kitchenpos.asis.domain.OrderTable;
 import kitchenpos.asis.domain.OrderType;
-import kitchenpos.product.domain.Product;
 
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.Arrays;
 import java.util.Random;
 import java.util.UUID;
 
-import static kitchenpos.product.ProductFixtures.product;
+import static kitchenpos.menu.MenuFixtures.menu;
 
 public class Fixtures {
     public static final UUID INVALID_ID = new UUID(0L, 0L);
-
-    public static Menu menu() {
-        return menu(19_000L, true, menuProduct());
-    }
-
-    public static Menu menu(final long price, final MenuProduct... menuProducts) {
-        return menu(price, false, menuProducts);
-    }
-
-    public static Menu menu(final long price, final boolean displayed, final MenuProduct... menuProducts) {
-        final Menu menu = new Menu();
-        menu.setId(UUID.randomUUID());
-        menu.setName("후라이드+후라이드");
-        menu.setPrice(BigDecimal.valueOf(price));
-        menu.setMenuGroup(menuGroup());
-        menu.setDisplayed(displayed);
-        menu.setMenuProducts(Arrays.asList(menuProducts));
-        return menu;
-    }
-
-    public static MenuGroup menuGroup() {
-        return menuGroup("두마리메뉴");
-    }
-
-    public static MenuGroup menuGroup(final String name) {
-        final MenuGroup menuGroup = new MenuGroup();
-        menuGroup.setId(UUID.randomUUID());
-        menuGroup.setName(name);
-        return menuGroup;
-    }
-
-    public static MenuProduct menuProduct() {
-        final MenuProduct menuProduct = new MenuProduct();
-        menuProduct.setSeq(new Random().nextLong());
-        menuProduct.setProduct(product());
-        menuProduct.setQuantity(2L);
-        return menuProduct;
-    }
-
-    public static MenuProduct menuProduct(final Product product, final long quantity) {
-        final MenuProduct menuProduct = new MenuProduct();
-        menuProduct.setSeq(new Random().nextLong());
-        menuProduct.setProduct(product);
-        menuProduct.setQuantity(quantity);
-        return menuProduct;
-    }
 
     public static Order order(final OrderStatus status, final String deliveryAddress) {
         final Order order = new Order();
