+++ src/test/java/kitchenpos/menu/MenuFixtures.java
@@ -1,0 +1,63 @@
+package kitchenpos.menu;
+
+import kitchenpos.menu.domain.Menu;
+import kitchenpos.menu.domain.MenuGroup;
+import kitchenpos.menu.domain.MenuProduct;
+import kitchenpos.product.domain.Product;
+
+import java.math.BigDecimal;
+import java.util.Arrays;
+import java.util.Random;
+import java.util.UUID;
+
+import static kitchenpos.product.ProductFixtures.product;
+
+public class MenuFixtures {
+
+    public static Menu menu() {
+        return menu(19_000L, true, menuProduct());
+    }
+
+    public static Menu menu(final long price, final MenuProduct... menuProducts) {
+        return menu(price, false, menuProducts);
+    }
+
+    public static Menu menu(final long price, final boolean displayed, final MenuProduct... menuProducts) {
+        final Menu menu = new Menu();
+        menu.setId(UUID.randomUUID());
+        menu.setName("후라이드+후라이드");
+        menu.setPrice(BigDecimal.valueOf(price));
+        menu.setMenuGroup(menuGroup());
+        menu.setDisplayed(displayed);
+        menu.setMenuProducts(Arrays.asList(menuProducts));
+        return menu;
+    }
+
+    public static MenuGroup menuGroup() {
+        return menuGroup("두마리메뉴");
+    }
+
+    public static MenuGroup menuGroup(final String name) {
+        final MenuGroup menuGroup = new MenuGroup();
+        menuGroup.setId(UUID.randomUUID());
+        menuGroup.setName(name);
+        return menuGroup;
+    }
+
+    public static MenuProduct menuProduct() {
+        final MenuProduct menuProduct = new MenuProduct();
+        menuProduct.setSeq(new Random().nextLong());
+        menuProduct.setProduct(product());
+        menuProduct.setQuantity(2L);
+        return menuProduct;
+    }
+
+    public static MenuProduct menuProduct(final Product product, final long quantity) {
+        final MenuProduct menuProduct = new MenuProduct();
+        menuProduct.setSeq(new Random().nextLong());
+        menuProduct.setProduct(product);
+        menuProduct.setQuantity(quantity);
+        return menuProduct;
+    }
+
+}
