+++ src/test/java/kitchenpos/product/ProductServiceTest.java
@@ -1,10 +1,10 @@
 package kitchenpos.product;
 
 import kitchenpos.asis.application.FakePurgomalumClient;
-import kitchenpos.asis.application.InMemoryMenuRepository;
+import kitchenpos.menu.InMemoryMenuRepository;
 import kitchenpos.product.application.ProductService;
-import kitchenpos.asis.domain.Menu;
-import kitchenpos.asis.domain.MenuRepository;
+import kitchenpos.menu.domain.Menu;
+import kitchenpos.menu.domain.MenuRepository;
 import kitchenpos.product.domain.Product;
 import kitchenpos.product.domain.ProductRepository;
 import kitchenpos.asis.infra.PurgomalumClient;
@@ -19,8 +19,8 @@ import java.math.BigDecimal;
 import java.util.List;
 import java.util.UUID;
 
-import static kitchenpos.Fixtures.menu;
-import static kitchenpos.Fixtures.menuProduct;
+import static kitchenpos.menu.MenuFixtures.menu;
+import static kitchenpos.menu.MenuFixtures.menuProduct;
 import static kitchenpos.product.ProductFixtures.product;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
